[BUGFIX] Metadata migration uses fields that don’t exist in TCA
Fields like sys_file:caption were removed from the table a while ago (in
EXT:media), and apparently have never been part of the core. To prevent
SQL errors when these fields are not present in a freshly migrated
installation, we now check for the existence of the fields in TCA (which
is a weak indicator that the field is present, but we can’t access the
database schema quite as easily).